// libs/stocks/enum/stocks.enum.ts
export enum Period {
    VIEW_VALUE_MAX = 'All available data',
    VIEW_VALUE_FIVE_YEAR = 'Five years',
    VIEW_VALUE_TWO_YEAR = 'Two years',
    VIEW_VALUE_ONE_YEAR = 'One year',
    VIEW_VALUE_YEAR_TO_DATE = 'Year-to-date',
    VIEW_VALUE_SIX_MONTHS = 'Six months',
    VIEW_VALUE_THREE_MONTHS = 'Three months',
    VIEW_VALUE_ONE_MONTH = 'One month'
  }

export enum PeriodValue {
    MAX = 'max',
    FIVE_YEAR = '5y',
    TWO_YEAR = '2y',
    ONE_YEAR = '1y',
    YEAR_TO_DATE = 'ytd',
    SIX_MONTHS = '6m',
    THREE_MONTHS = '3m',
    ONE_MONTH = '1m'
}